feat(donation): mobile view and make card payment default
#95

- refactor donation/donation success page
- split out into hooks and library functions
- add a mock for stripe for local development
- enhance mobile/nonmobile view for header
