+++ components/page/header.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import BackButton from '@/components/back-button';
 import { ReactNode } from 'react';
 import { Badge } from '@/components/ui';
@@ -14,6 +15,47 @@ export function PageHeader({
   tags?: string[];
   children?: ReactNode;
 }) {
+  const headLine = useMemo(() => {
+    const items = [];
+    if (typeof message === 'string') {
+      items.push(
+        <div key="message" className="text-sm text-muted-foreground">
+          {message}
+        </div>,
+      );
+    }
+    items.push(
+      <h1 key="title" className="text-lg font-semibold sm:text-4xl">
+        {title}
+      </h1>,
+    );
+    return items;
+  }, [message, title]);
+  const navLine = useMemo(() => {
+    const items = [];
+    if (featured) {
+      items.push(
+        <Badge key="featured" variant="secondary" className="px-3 py-1">
+          Featured
+        </Badge>,
+      );
+    }
+    if (Array.isArray(tags) && tags.length) {
+      items.push(
+        <div key="tags" className="flex flex-wrap">
+          {tags?.map((tag) => (
+            <span key={tag} className="pr-1 text-sm text-gray-500">
+              {tag}
+            </span>
+          ))}
+        </div>,
+      );
+    }
+    if (items.length === 0) {
+      items.push(headLine);
+    }
+    return items;
+  }, [featured, tags, headLine]);
   return (
     <header className="mx-auto inline items-center justify-between p-8">
       {/* campaigns[slug] ^px-4 py-8 */}
@@ -22,26 +64,12 @@ export function PageHeader({
         {/* campaigns[slug] <div className="flex items-center gap-2 py-1"> */}
         <div className="mb-8 flex flex-row items-center gap-2 py-1 md:mb-2 md:ml-4">
           <BackButton />
-          {featured && (
-            <Badge variant="secondary" className="px-3 py-1">
-              Featured
-            </Badge>
-          )}
-          <div className="flex flex-wrap">
-            {tags?.map((tag) => (
-              <span key={tag} className="pr-1 text-sm text-gray-500">
-                {tag}
-              </span>
-            ))}
-          </div>
+          {navLine}
         </div>
       </div>
       <div>
         <div className="ml-2 flex flex-col md:ml-4 md:content-start">
-          {typeof message === 'string' && (
-            <div className="text-sm text-muted-foreground">{message}</div>
-          )}
-          <h1 className="text-lg font-semibold sm:text-4xl">{title}</h1>
+          {navLine[0] !== headLine && headLine}
           {children ? (
             <div className="flex items-center py-2">{children}</div>
           ) : null}
