+++ components/payment/status-loading.tsx
@@ -1,6 +1,6 @@
 export function PaymentStatusLoading() {
   return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50/50 p-4">
+    <div className="flex items-start justify-center bg-gray-50/50 p-4">
       <div className="text-center">
         <div className="mx-auto h-8 w-8 animate-spin rounded-full border-b-2 border-gray-900"></div>
         <p className="mt-4">Loading payment status...</p>
