+++ components/payment/switch-wallet-network.tsx
@@ -1,0 +1,44 @@
+import { Button } from '@/components/ui';
+import { useCallback } from 'react';
+import { useToast } from '@/hooks/use-toast';
+import { useNetworkCheck } from '@/hooks/use-network';
+import { AlertCircle } from 'lucide-react';
+export function PaymentSwitchWalletNetwork() {
+  const { toast } = useToast();
+  const { isCorrectNetwork, switchToAlfajores } = useNetworkCheck();
+
+  const handleNetworkSwitch = useCallback(async () => {
+    try {
+      await switchToAlfajores();
+    } catch (error) {
+      console.error('Error switching network:', error);
+      toast({
+        title: 'Network Switch Failed',
+        description: (
+          <div className="flex items-center gap-2">
+            <AlertCircle className="h-4 w-4 text-destructive" />
+            <p className="text-sm font-medium">
+              Unable to switch network automatically. Please switch to Celo
+              Alfajores manually in your wallet.
+            </p>
+          </div>
+        ),
+        variant: 'destructive',
+        duration: 5000,
+      });
+    }
+  }, [switchToAlfajores, toast]);
+  if (isCorrectNetwork) {
+    return null;
+  }
+  return (
+    <Button
+      variant="secondary"
+      size="sm"
+      onClick={handleNetworkSwitch}
+      className="w-full text-black"
+    >
+      Switch Network
+    </Button>
+  );
+}
